+++ utils/light/ltface.c
@@ -55,6 +55,7 @@ static double CastRay (vec3_t p1, vec3_t p2)
 
 	if (t == 0)
 		t = 1;		// don't blow up...
+
 	return sqrt(t);
 }
 
@@ -203,10 +204,10 @@ static void CalcFaceExtents (lightinfo_t *l)
 
 		for (j = 0 ; j < 2 ; j++)
 		{
-			val = v->point[0] * tex->vecs[j][0] + 
-				v->point[1] * tex->vecs[j][1] +
-				v->point[2] * tex->vecs[j][2] +
-				tex->vecs[j][3];
+			val =	((double)v->point[0] * (double)tex->vecs[j][0]) +
+				((double)v->point[1] * (double)tex->vecs[j][1]) +
+				((double)v->point[2] * (double)tex->vecs[j][2]) +
+				(double)tex->vecs[j][3];
 			if (val < mins[j])
 				mins[j] = val;
 			if (val > maxs[j])
